ci/test: use cargo-deny to check licenses and duplicate dependencies
Since Rust crates tend to be tiny, it's very easy to accidentally bring
in a ton of duplicate transitive dependencies. Use cargo-deny to lint
against this happening. When duplicate dependencies are unavoidable,
they can be explicitly allowed in the deny.toml config file.

As an added bonus, cargo-deny will also prevent us from depending on
non-OSS code.

This commit also upgrades us to Rust 1.41, since cargo-deny depends on
features in Rust 1.41. The code changes are the result of changes in
Rust 1.41's rustfmt.