billing-demo: remove dependency on uuid_b64
uuid_b64 brings in old duplicate versions of a number of dependencies
(base64, uuid, lazy_static), as well as a few dependencies not used at
all elsewhere (error-chain, inlinable_string).

It doesn't seem all the important to the functioning of the billing-demo
to use a base64-encoded string, so just drop it. (Ultimately wouldn't we
want to use a u128 instead of a string?)