expr,sql: support distinct explicit and implicit casts
Casting between certain data types has different behavior depending on
whether the cast is explicitly requested by the user (i.e., via CAST or
the :: operator) or implicitly selected by a function or operator.
Generalize that notion and thread it through our sql and expr crates.

So far this only impacts the `concat` function, but having it as a
generalizable form of cast behavior paves the way for the other data
types that have different explicit and implicit casting behavior.