dataflow,sql: ensure kafka offset column is accounted for
Raw Kafka sources were not properly reporting their mz_offset column.
The situation is weird because Avro and Protobuf formats do not
propagate this column, while regex and CSV formats do. Just hack around
the issue for now, and add additional tests to decrease the probability
of the type descriptor skewing from the actual result set in the future.

This issue was discovered by @JLDLaughlin while working on #2259.