comm: fix flakes in test_pingpong
When a comm channel closes, the switchboard attempts to recycle the
underlying `TCPStream` for use by a future comm channel. Recycling
involves calling `peer_addr()` on this `TCPStream` so that the
connection can be sorted into the correct pool.

If the peer closes the connection before the call to `peer_addr()`, the
call to `peer_addr()` will fail with an error like "transport endpoint
is not connected." We would previously panic upon receiving this error,
which caused occasional flakes in test_pingpong in CI (see #2291).

This wasn't a problem in actual deployments of Materialize, as all
workers are long-lived and no peers ever attempt to terminate
connections.

This patch allows the comm::Connection::addr method to return an error,
and the connection pool simply declines to recycle a connection if
retriving its peer address results in an error.

Fix #2291.