expr: compute output type of coalesce function more obviously
The output type computation for the coalesce function was previously
quite clever and would mark coalesce as non-nullable if all of its
inputs were non-nullable. It also had an unnecessary check for
ScalarType::Unknown: all inputs to coalesce are made to have the same
type, so there is no need to filter out ScalarType::Unknown when
computing a coalesce invocation's output type.

It is quite a bit clearer if we instead declare the output type of
coalesce to be:

  - unconditionally nullable
  - the same scalar type as its inputs (which are required to all be
    of the same type)

We're not losing an optimization here, either. If a coalesce is indeed
applied to all non-nullable arguments, the new support for reducing
coalesce expressions (see #2298) will remove the coalesce entirely, and
so the non-nullability of the first input will be accessible to the
optimizer in that way.