Update the `since` field of IndexState. (#2336)
This field was not previously being updated, which made validity tests against it vacuously true. This meant that races between compaction and upper frontiers could result in incorrect results. There should now be an error instead (the error has always existed, but the state on which it was based was not correct).