testdrive: fix kafka-verify
A Kafka producer locally caches metadata about what topics and
partitions are available. Attempting to produce to a partition that is
not present in the local cache will result in an UnknownPartition error.

The default topic metadata refresh interval is 5m, which means it could
take a full five minutes between when testdrive creates a new partition
and when it can produce to that partition. This is obviously untenable,
so the refresh interval was decreased to 10ms in #2214.

Unfortunately, this had the side-effect of wedging any Kafka consumers.
Their metadata would always be out of date, and so they'd spend all
their time updating their metadata and never yield any records.

This patch restores topic.metadata.refresh.interval to its default,
fixing kafka-verify and restoring the test that used kafka-verify. Then,
to ensure that the Kafka producer has sufficiently fresh metadata, it
adjusts kafka-create-topic and kafka-add-partitions to forcibly refresh
the producer's metadata after the topic or partitions are created.

To prevent similar disasters in the future, there is no longer a Kafka
consumer that is shared across all the actions in a test file. Actions
that need a Kafka consumer (like kafka-verify) create their own on
demand. All metadata fetches will thus be forced through the Kafka
producer, where they will ensure the cache is up to date.

Fix #2299.