Allow back-references in Avro schemas, and rework schema resolution. (#2375)
This will substantially improve our ability to recognize and support valid Avro schemas, which the `munge` function was attempting to do before, but imperfectly. Essentially, the issue was that the `avro-rs` didn't support backreferences (reference by name to previously-defined pieces of the schema).

To accomplish this, nodes in the schema tree can now be references to the index of other nodes in a schema-wide list of named pieces. Thus functions dealing with schemas need to become stateful - a local node in the tree no longer carries all the information needed to interpret it (since it needs to refer to a schema-wide mapping of names to sub-schemas).

Thus, we introduce a new abstraction: `SchemaNode`, which carries both a reference to a chunk of the schema, and a reference to the schema-wide state. Functions that formerly recursively traversed schemas have been rewritten to use `SchemaNode` (or its closely-related cousin, `SchemaNodeOrNamed`, which refers to either a chunk of the schema *or* a not-yet-looked-up named reference).

This PR also fixes unions to allow multiple named variants of the same kind (so long as they have different names), and completely rewrites the schema resolution logic. Previously, we were resolving a *value* against a *schema*, which is not actually possible to do in general: values don't carry the names of the types they instantiate, so when resolving (for example) a record value against a union that contains multiple records with different names, the code can't know which to attempt to resolve against.

Now, instead of resolving values against schemas, we resolve *schemas* against schemas, recording, at each node, all the information necessary to parse a value written by the writer and return a value understandable by the reader.

I have also deleted a bunch of code in `avro-rs` that we never plan to use, like Serde integration.