Allow ScalarExpr in Join keys (#2235)
This PR means to allow general ScalarExprs in join keys, rather than just columns. There are several cases where this helps, most prominently when there is a single constant value to be joined with other values (e.g. a filter to implement as an indexed look-up). This shows up in @wangandi's work, but also in set intersection work, plausibly.

Anyhow, it also just seems like the right thing to do, and simplifies logic in many cases. In particular, it may obviate the need for simplify.rs which currently exists only to work around this limitation.