coord,sql: make sink topic creation asynchronous
Just like sources need an asynchronous pre-planning step
("purification"), sinks need an asynchronous post-planning step. With
Kafka sinks, for example, we need to create the topic the sink will
write to.

Previously we'd do that by blocking the SQL planning thread, but that
has the usual performance caveats. This patch teaches the coordinator to
instead create the topic asynchronously. The plumbing is a real drag, as
we write a state machine by hand that's split across a few crates:

  * The SQL layer creates a "sink connector builder" that contains all
    the necessary pieces of state to actually build a sink connector.

  * The coordinator installs a temporary catalog item to reserve the
    name while the connector is built. Because we can't transactionally
    create a Kafka topic, we need to "lock" the name while we go create
    the Kafka topic, otherwise another session might claim the name.
    (We cannot fail for any reason after creating the Kafka topic or
    we'll leak the topic.)

  * The coordinator launches an asynchronous task to create the topic.

  * After the topic is created, the future resolves with the actual
    connector, and the coordinator updates the catalog with the actual
    connector.

  * The dataflow layer builds the sink using the connector.