ci/test: move security advisories check to dedicated job
We don't want historical builds or PR builds failing just because a
security vulnerability was discovered in one of our dependencies. These
are serious, but not serious enough to warrant nondeterminism in our CI
pipeline.

Instead introduce a separate CI job whose only responsibility is to
check for these vulnerabilites in the Rust advisory database. That job
will be scheduled to run daily and post failures in Slack.