doc/developer: tweak formatting and hierarchy
Enforce a bit more of a consistent style on the developer docs. Make the
README a bit easier on the eyes with a more spaced out TOC, and also add
a few more pointers for new contributors. Finally, group together the
docs that form the "developer guide", which are meant to be consumed
together, rather than as standlone reference material.