Remove uses of un-demanded columns (#2603)
Demand analysis would indicate that some columns could be set to dummy values, but eager evaluation in e.g. Map and Reduce could result in runtime errors. This is not a problem at the moment in the code as runtime errors are just cast to Null and never read. However, with upcoming runtime errors it seemed worth fixing anyhow.

fixes #2601