test: add some JavaScript tests
The rust-postgres client is a bit too helpful in doing client-side
validations, making it difficult to test things like a query with an
incorrect number of bind parameters. The JavaScript Postgres driver is
far less strict about this, so add support for writing tests in
JavaScript.

We'll likely want to expand our support for testing in other languages,
so I've set up a new test/lang directory to facilitate future expansion
of these non-Rust test suites.

Fix #2610.