Support more SHOW ... WHERE
Fixes #1529.

The only remaining unsupported case is SHOW VIEWS/SOURCES, because we
don't plumb down the needed info via the catalog.

I got rid of some ordering stuff, I can add it back in the form of a
RowSetFinishing, but it doesn't seem to have been tested so I'm not sure
it was actually important.