mzbuild: `git clean` with `-d` flag
In newer versions of git the `-d` flag is redundant, because we've
specified an explicit path to clean, but in old versions of Git, like
the one we have in the CI builder, the `-d` flag needs to be passed
explicitly to get `git clean` to remove ignored directories. [0]

This was causing mzbuild failures in CI, where the ci-cargo-test image
could have build directories lying around from the last job to run on
the agent.

[0]: https://github.com/git/git/commit/e86bbcf987faad8e487beb814351a404fa80957f