Introduce AstDisplay
This commit introduces a new trait for formatting ASTs that is distinct
from fmt::Display. The main difference is that it is orchestrated by a
Formatter we control, and can thus be configured to print the entire
tree in specific ways. It comes with the unfortunate ergonomic downside
of no longer being able to use the format!/write! macro. Suggestions for
how to improve the ergonomics here are welcome!

This is specifically to enable #2667, but I foresee being able to
configure the formatting of an AST to be useful in the future as we use
them for more things (in the past we used this kind of infrastructure
for things like forcing full-qualification of names)..