testdrive: disable automatic topic creation
Testdrive and Materialize often race to create the same topic. We want
testdrive to win, as it issues actual create topic requests with
specific partition counts, while Materialize is accidentally creating
the topic via Kafka's auto-topic creation feature. Auto-topic creation
is usually disabled in production, so it's actually more realistic to
test with it off, anyway.

This exposed we were forgetting to create topics in a few testdrive
scripts.

The hypothesis is that this is a cause of one the flakes described in