ci: ensure all mzbuild images are pushed
If pushing an image fails (e.g. because the Docker Hub permissions are
incorrect), that push won't be retried on subsequent CI builds, because
the image will be available locally and the push will get skipped.

Rework the build jobs so that all images are checked for existence on
Docker Hub and pushed if they do not exist.