Extract arrangement frontier state (#2826)
This PR extracts from the coordinator logic related to managing the frontier state of arrangements. These methods were previously member methods of the coordinator, rather than restricted to the index they act on. This is part of a progressive thinning out of tasks from the coordinator and in to comprehensible modules.