Fix TAIL's use of TIMESTAMP{,TZ} as timestamps
This was inadvertedly using seconds since epoch instead of milliseconds.
This commit changes it to use the milliseconds.

I'd like to add a test for this, but I'm still not sure how to do so in
a nice way, so I'll just include the promise that I tested this out by
hand (thought I did last time, I guess I confused myself somehow),
unless there are better ideas!