Fix panic in demand analysis
Previously, we would inform the input to a FlatMapUnary that we
expected it to produce even the column we ourselves were producing,
which understandably caused it anguish. This change only asks for the
columns the input is capable of producing. I couldn't find an existing
place that set-returning functions were tested, so I added an SLT file.