Introduce `transform` crate (#2911)
This PR moves transformation logic out of the expr crate and in to its own crate. The rationale is that expr is a very popular crate, but many of its users do not depend on the state of transformations. In particular, the dataflow crate is expensive to rebuild and while it does depend on expr it does not depend on transformations.

This PR also removes three transformations that are no longer in use, and flags two for further review. It also slaps up documentation everywhere. I'd love to require doc code examples, but the lint (missing-doc-code-examples) doesn't seem to have an effect that I could tell.