Re-organize render (part 1) (#2916)
The first part of attempts to smash up render/mod.rs into smaller parts. These are the easier parts that were already in their own methods.

One relevant change is that the methods render_foo should be called only after ensuring that their inputs are rendered. This is because the methods can only make that recursive call back in to the Context if they have the most general type parameters, which is more restrictives than we actually require. The change allows for implementations that are more clearly Timestamp-independent, and also a little less shuffling around of things like scopes and worker ids (important for some operators, not for others).