sql: introduce a catalog trait
The dependency graph is presently

    catalog <- SQL <- coord
       ^               |
       |               |
       +---------------+

because the SQL planner needs the catalog to perform name resolution,
but the coordinator needs the catalog to actually persist state to disk.
The catalog really wants to be a part of the coordinator so it can be
responsible for storing all of the coordinator's metadata, but this
results in a circular dependency.

Prepare to break that circular dependency by introducing a "PlanCatalog"
trait that defines the interface that the SQL package needs from a
catalog. This trait is presently quite ugly, to facilitate the refactor,
but will hopefully simplify over time.