Document decorrelation some (#2935)
This PR introduces a bunch of explanatory text for parts of decorrelation, written as part of understanding what it does. I hope they are mostly correct. I stopped at ScalarExpr::applied_to but I'm happy to pick that up.

I put decorrelation methods in a module, with some module-level documentation.

I removed the Result wrappers from the type; the only potential error was the ensure! macro which I replaced with a panic! (perhaps incorrectly).