doc/user: fix Avro OCF source documentation
Avro OCF sources are a bit special and are a connector, rather than a
format. Correct the docs accordingly.

Also remove a bit of magic in the rendering of the CREATE SOURCE
introduction that seemed to be resulting in a lot of incorrect or
awkward constructions ("connect Materialize to Avro-formatted Kafka
topic", "Avro Object Container File-formatted local files").

Fix #2969.