doc/user: rework css
This reworks the docs site CSS so that it is not dependent on either
Bootstrap or the marketing website CSS. This should improve its
maintainability, despite the small amount of required duplication, as we
don't need to constantly fight against rules from other stylesheets.

I also took the opportunity to clean up some spacing, so that the
information hierarchy is better reflected in the visual whitespace, and
to standardize colors based on the official Materialize style guide.

New features include a scrollspy plugin that automatically highlights
TOC headings in the sidebar as you scroll, and the ability to use
non-fancy ordered lists when the list elements do not contain paragraph
elements.